requirements.txt
streamlit
pandas
joblib
python-dotenv
google-genai
